requirements.txt
Flask
psycopg2-binary
flask-sqlalchemy
flask-migrate